+++ packages/bridge-controller/src/constants/bridge.ts
@@ -91,6 +91,7 @@ export const DEFAULT_BRIDGE_CONTROLLER_STATE: BridgeControllerState = {
   assetExchangeRates: {},
   minimumBalanceForRentExemptionInLamports: '0',
   tokenWarnings: [],
+  tokenSecurityTypeDestination: null,
   quoteStreamComplete: null,
 };